Fix IAM policy naming for inline policies with empty string Sid field. (#1389)
### Summary
> AWS is breaking its own rules for some roles it has created in
IdentityCenter. IdentityCenter created roles with inline policies that
can have multiple statements, all with the Sid of an empty string. This
means cartography will only load the last statement as they will all
have the same AWSPolicyStatement.id. This change checks if the Sid is an
empty string and uses the index numbering if it is encountered.



### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/...


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [ ] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

Co-authored-by: = <=>
Co-authored-by: Alex Chantavy <achantavy@lyft.com>