Github "Immediate" User Team Membership (#1395)
### Summary

This PR adds to the Github graph, adding user membership in teams, for
users who are 'immediate' members of a team.

In case it is unclear or for people newer to Github, note: this is
focusing on
['immediate'](https://docs.github.com/en/graphql/reference/enums#teammembershiptype)
membership to a team, meaning a member is in the team directly as
opposed to being in a child team. A user could be considered a member of
a team if they are members of a child team, but this PR maps only
'immediate' membership. (In a follow-up PR we'd like to add child teams
to the graph, which we think will complete the membership picture.)

We think this is a valuable addition to the graph because our broad
intent is to understand all access a user has, and (at least in our org)
most access to repos is granted via team. If we do not know who is in
the team, then, we do not know who has access.

#### Illustration of the intention
![Cartography AMPS User Direct Team
Membership](https://github.com/user-attachments/assets/7d3d70ab-ab16-4a21-8970-3f9d2b8fe525)

#### Screencaps

**EXAMPLE USER LOOKUP**

BEFORE
(empty result because nothing exists)
![Screenshot 2024-12-03 at 5 07
00 PM](https://github.com/user-attachments/assets/908e8e96-179d-494a-ac7d-acc03ee54ab1)

AFTER
![Screenshot 2024-12-03 at 5 06
04 PM](https://github.com/user-attachments/assets/073376ee-b8d6-4b68-8d7e-246e9f9601a2)

**OVERVIEW OF COUNTS OF EACH TYPE**

BEFORE
(empty result because nothing exists)
![Screenshot 2024-12-03 at 5 09
03 PM](https://github.com/user-attachments/assets/c4758f8f-71ec-49b5-a94a-c3d464a9a51e)

AFTER
![Screenshot 2024-12-03 at 5 08
44 PM](https://github.com/user-attachments/assets/cc0be586-c1bf-440b-a8a0-6bc236272509)


### Related issues or links

None


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

**N/A** If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Daniel Brauer <dbrauer@etsy.com>