GitHub: introduce ADMIN_OF relationship between users and orgs (#1403)
### Summary

This PR fixes the issue described in issue #1374—please see the issue
and discussion there for details, but here it is in summary:

The 'role' property on a GitHubUser node expresses whether a user is an
'ADMIN' (aka Owner) or a 'MEMBER' of a GitHubOrganization. For a graph
with multiple orgs, however, the role property will reflect only one of
the user-org relationships. If the user is an admin in one org and a
member in another, for example, the 'role' property will be wrong for
one of those relationships. To fix this, this PR shifts the role to be
expressed in the relationship between GitHubUser and GitHubOrganization.

Below are a few before/after screen caps to show intent and the PR
working.

#### **USER IS AN OWNER OF ONE ORG, MEMBER OF ANOTHER**

BEFORE
Here the user is graphed as 'MEMBER_OF' to both orgs. The 'role'
property (not pictured) is set to reflect one of those relationships, so
it is partly wrong.
![Screenshot 2024-12-09 at 5 22
15 PM](https://github.com/user-attachments/assets/4b1094ba-0020-4d10-a397-52653e0214e9)

AFTER
The relationship is now updated to reflect that the user is an admin of
one of the orgs. And, as suggested by @achantavy in #1374, admins now
get two user-org relationships "so that we can determine what org a user
belongs to in a uniform way".
![Screenshot 2024-12-09 at 5 08
04 PM](https://github.com/user-attachments/assets/6c654b5f-0363-4045-9167-2f6459371cbc)

#### **MEMBERS AND UNAFFILIATEDS REMAIN THE SAME**

BEFORE & AFTER
This user is handy for demonstration purposes because they are a member
of one org, and unaffiliated to another (the user in question is an
enterprise owner, and is not a member of the second org). These
relationships remain untouched by this PR, and so the before and after
results for the query shown are identical.
![Screenshot 2024-12-09 at 5 07
20 PM](https://github.com/user-attachments/assets/4dffb9a5-691d-4099-9169-9f4195814c7b)


#### **RELATIONSHIP COUNTS**

BEFORE
Only MEMBER_OF and UNAFFILIATED is graphed, and to know who is an ADMIN
you have to look at a property:
![Screenshot 2024-12-11 at 8 44
21 PM](https://github.com/user-attachments/assets/ec8c7675-13d1-4ecd-a059-2154a3e9beb1)
![Screenshot 2024-12-11 at 9 03
52 PM](https://github.com/user-attachments/assets/c78d5db8-7ff4-44b1-b76d-67ab6bd5011f)
(Note there are 16 admins, but also note how I had to specify folks who
are 'UNAFFILIATED' in my query. If I hadn't done that, I would've gotten
a higher number, as some of those folks are admins but of a different
org. So the result would have been wrong.)

AFTER
In addition to the exiting relationships that were there before, there
are 16 new ADMIN_OF relationships for every node that _had_ an ADMIN
role property.
![Screenshot 2024-12-11 at 8 41
56 PM](https://github.com/user-attachments/assets/d70a5583-9366-42f9-82b7-5c51d426a7f8)



### Related issues or links

- #1374 


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

**NOT APPLICABLE** you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

Signed-off-by: Daniel Brauer <dbrauer@etsy.com>