#1413: Support automatic cleanups for NodeSchemas without a sub resource (#1419)
### Summary
> Describe your changes.

#1413 

Removes a restriction so that cartography's data model can now
automatically clean up NodeSchemas that don't have a tenant
relationship.

In these cases, the nodes will not be deleted even if they are
considered stale, but their stale relationships will be.

We do it this way because cartography syncs assets one tenant at a time,
and stale nodes can only be safely deleted if they are tied to a tenant
so that we do not erroneously delete nodes attached to other tenants.

### Related issues or links
> Include links to relevant issues or other pages.
- https://github.com/lyft/cartography/issues/1413

We discussed this in the above issue but I'll summarize a bit here: in
the case of GitHub, GitHubUsers don't have a true "tenant" relationship
with their organization because they can exist independent of their
organizations. In this case, it makes sense for a GitHubUser to have a
node schema that does not have a tenant rel. If we want to list out all
of the users in an organization, then we simply query for the
organization and its attached users.


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

---------

Signed-off-by: Alex Chantavy <chantavy@gmail.com>