Fix #386: avoid duplicate values in exposed_internet_type for AWS EC2 and GCP (#1425)
### Summary
> Describe your changes.

Fixes bugs where we added duplicate values to the
`exposed_internet_type` field in the AWS EC2 and GCP analysis jobs.

### Bugfix 1: Added an additional CASE condition to handle existing
lists:
- Add `['direct']` (as an array) instead of just the string 'direct' so
that the concatenation in Neo4j works as expected
- Added an ELSE clause to return the existing list unchanged if 'direct'
is already present
This now properly maintains a list without duplicating the 'direct'
entry.


### Bugfix 2. Remove a slow AllNodesScan
The query
```cypher
MATCH (n)
WHERE 
    n.exposed_internet IS NOT NULL AND 
    labels(n) IN ['AutoScalingGroup', 'EC2Instance', 'LoadBalancer', 'LoadBalancerV2'] 
WITH n LIMIT $LIMIT_SIZE 
REMOVE n.exposed_internet, n.exposed_internet_type 
return COUNT(*) as TotalCompleted
```

is supposed to remove the `exposed_internet` and `exposed_internet_type`
fields from all nodes `n` if node `n` is an AutoScalingGroup,
EC2Instance, LoadBalancer, or LoadBalancerV2. However this query is
wrong because `labels(n)` will return `['EC2Instance']`, and
`['EC2Instance']` is not in the array `['AutoScalingGroup',
'EC2Instance', 'LoadBalancer', 'LoadBalancerV2']`.

Further, the larger issue is that it would be more performant to match
on specific labels instead of all nodes for a large graph. To fix we
will match on specific labels.


## Before

![image](https://github.com/user-attachments/assets/87ae742e-a3d4-4f40-aaf9-4fc953dfb7a2)

The query removes nothing, and `exposed_internet_type` is a list of
repeating items:


![image](https://github.com/user-attachments/assets/9e80d8f6-cc92-4939-948a-48f548d80557)


## After

![image](https://github.com/user-attachments/assets/373707d0-3373-40b6-bccd-96998514524e)

Now that I've separated it to specific labels, the query performs as
intended.


![image](https://github.com/user-attachments/assets/ce0bba63-9b7e-44a7-ba8f-5cfdef6b4e02)


### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/386







### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [ ] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

---------

Signed-off-by: Alex Chantavy <chantavy@gmail.com>